// src/app/pages/brewery/model/brewery.model.ts
export interface IBrewery {

  "id": "string",
  "name": "string",
  "brewery_type": "string",
  "address_1": "string",
  "address_2": "string",
  "address_3": "string",
  "city": "Norman",
  "state_province": "string",
  "postal_code": "number",
  "country": "string",
  "longitude": "number",
  "latitude": "number",
  "phone": "4058160490number",
  "website_url": "string",
  "state": "string",
  "street": "string";

}

export class Brewery implements IBrewery {
  "id": "string";
  "name": "string";
  "brewery_type": "string";
  "address_1": "string";
  "address_2": "string";
  "address_3": "string";
  "city": "Norman";
  "state_province": "string";
  "postal_code": "number";
  "country": "string";
  "longitude": "number";
  "latitude": "number";
  "phone": "4058160490number";
  "website_url": "string";
  "state": "string";
  "street": "string";

}